pyproject.toml
[tool.poetry]
name = "ailiga"
version = "0.1.0"
description = ""
authors = ["Alexander Puck Neuwirth <alexander@neuwirth-informatik.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
SuperSuit = "*"
torch = "1.13.0"
tianshou = "0.4.10"
#tianshou = {git = "https://github.com/thu-ml/tianshou", rev= "06aaad460e402bfa4ae3b631a5e7f64d0b1d5405"}
#tianshou = {path = "/home/apn/git/tianshou/"}
PettingZoo = "<1.22.0"
tqdm = "*"
pqdm = "*"
#stable-baselines3 = "1.6.2"
#stable-baselines3 = {git ="https://github.com/carlosluis/stable-baselines3.git", branch="fix_tests", extras= ["extra"]}
pyglet = "1.5.27"
cfg_load  = "*"
rlcard = "*"
smpl_doc =">=1.1.1"
smpl_io ="*"


#gym = "<=0.21"
#gym = "*"
#ray = { version = "2" , extras = ["rllib"]}
#open_spiel = "*"

[tool.poetry.scripts]
ailiga-train = "ailiga.train:main"
ailiga-tournament = "ailiga.tournament:main"


[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "*"
sphinx-rtd-theme = "*"
sphinxcontrib-napoleon = "*"
sphinx-autoapi = "*"
sphinx_autobuild = "*"
pandoc = "*"
numpydoc = "*"
myst-parser  = "*"
toml = "*"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
